update to java 17, initial linter implementation

+++ src/main/java/pl/asie/libzxt/ZxtExtensionBlock.java
@@ -39,7 +39,7 @@ public class ZxtExtensionBlock {
         this(flags, id, null);
     }
 
-    public short getFlags() {
+    public int getFlags() {
         return flags;
     }
